reorganizes the code, introduces preference slice
