--- frontend/src/hooks.ts
@@ -1,22 +1,0 @@
-import { useState, useEffect } from 'react';
-
-export function useMediaQuery(query: string): boolean {
-  const [matches, setMatches] = useState(false);
-
-  useEffect(() => {
-    const media = window.matchMedia(query);
-
-    if (media.matches !== matches) {
-      setMatches(media.matches);
-    }
-
-    const listener = () => {
-      setMatches(media.matches);
-    };
-
-    media.addEventListener('resize', listener);
-    return () => media.removeEventListener('resize', listener);
-  }, [matches, query]);
-
-  return matches;
-}
